Lets go, breaking code into parts.

+++ libft/ft_cpyarray.c
@@ -6,15 +6,15 @@
 /*   By: tsharma <tsharma@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/04 18:27:52 by tsharma           #+#    #+#             */
-/*   Updated: 2022/08/04 18:29:34 by tsharma          ###   ########.fr       */
+/*   Updated: 2022/08/14 16:03:23 by tsharma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	*ft_cpyarray(int *dst, int *src, size_t n)
+int	*ft_cpyarray(int *dst, int *src, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (src == NULL && dst == NULL)
